requirements.txt
requests==2.32.5
python-dotenv==1.2.1
beautifulsoup4==4.14.3
flask==3.1.3
